Fix loot tables for armor
Second part of loot table revamp, featuring armor item slots. The one
exception is all the crafted legendary helms, these will be done in a
separate commit as they require a lot more work by creating an
invdividual loot table for each one and editing the blueprints to point
to this loot table instead of the item itself.