Turn off hiding prefixes/suffixes
GQ items are copypaste is some areas such as this, changing them so that
affix'd versions in DAIL show the affix names so people don't come
crying. Did not do this to staves yet because they don't even drop or
are finished so don't want to potentially cause a conflict with a GQ
update.