My take on some changes to Mystic Warrior
Changed the debuffs on the base Crippling Curse to negative OA stats to
stack better with other debuffs and be more useful against major
enemies. Removed the petrify on the 2nd modifier for negative DA for the
same reason. Also tuned up the life leech on it and tuned down the
energy leech on the 1st modifier of Wracking Aura, you leeched more of
the latter than the former despite the size of their respective pools.
Also fixed some minor bugs and typo stuff.